Consolidate RaidInstanceIndexItem and DungeonInstanceIndexItem into InstanceIndexItem
