+++ frontend/src/components/ExpansionDropdown.js
@@ -1,15 +1,12 @@
 import React, { useState } from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faCaretDown, faCaretRight } from '@fortawesome/free-solid-svg-icons'
-import RaidInstanceIndexItem from './RaidInstanceIndexItem';
-import DungeonInstanceIndexItem from './DungeonInstanceIndexItem';
+import InstanceIndexItem from './InstanceIndexItem';
 
 function ExpansionDropdown({expansion, expansionName, type}) {
 	const [ isOpen, setIsOpen ] = useState(false);
 
   const caret = isOpen ? faCaretDown : faCaretRight;
-
-  const InstanceIndexItem = type === 'raid' ? RaidInstanceIndexItem : DungeonInstanceIndexItem;
 
 	return (
     <li className={'index-item border'.concat(isOpen ? ' open' : '')}>
@@ -21,11 +18,9 @@ function ExpansionDropdown({expansion, expansionName, type}) {
         { isOpen && ( expansion.instances ? 
           <div className='expansion-instance-container'>
             { 
-              Object.entries(expansion.instances).map(([instanceName, instance]) => {
-                return (
-                  <InstanceIndexItem key={instanceName} instance={instance}/>
-                )
-              })
+              Object.entries(expansion.instances).map(([instanceName, instance]) => (
+                  <InstanceIndexItem key={instanceName} instance={instance} type={type}/>
+              ))
             }
           </div> 
           :
